Extract "ToolCalls" to own file.
