+++ pylib-v1/ailets/models/gpt4o/lib/sse.py
@@ -7,6 +7,8 @@ from ailets.cons.util import write_all
 
 from typing import TypedDict, NotRequired
 
+from ailets.models.gpt4o.lib.tool_calls import ToolCalls
+
 
 class Delta(TypedDict):
     role: NotRequired[str]
@@ -43,45 +45,14 @@ def escape_json_value(s: str) -> str:
     return "".join(result)
 
 
-class ToolCalls:
-    def __init__(self, tool_calls: Optional[Sequence[Mapping[str, Any]]]) -> None:
-        self.tool_calls: list[ContentItemFunction] = []
-        if tool_calls is None:
-            return
-        for tool_call in tool_calls:
-            assert tool_call["index"] == len(
-                self.tool_calls
-            ), "Tool call indices must be sequential"
-            assert is_content_item_function(tool_call), "Tool call must be a function"
-            self.tool_calls.append(tool_call)
-
-    def delta(self, tool_calls: Optional[Sequence[Mapping[str, Any]]]) -> None:
-        if tool_calls is None:
-            return
-        for tool_call in tool_calls:
-            index = tool_call["index"]
-            if index < 0 or index >= len(self.tool_calls):
-                raise ValueError(f"Tool call index {index} is out of range")
-            base_tool_call = self.tool_calls[index]
-            assert "function" in tool_call, "Tool call must have 'function' key"
-            function = tool_call["function"]
-            assert isinstance(function, dict), "'function' must be a dictionary"
-            assert list(function.keys()) == [
-                "arguments"
-            ], "'function' must only have 'arguments' key"
-
-            base_tool_call["function"]["arguments"] = function["arguments"]
-
-    def get_tool_calls(self) -> list[ContentItemFunction]:
-        return self.tool_calls
-
 
 class SseHandler:
     def __init__(self, runtime: INodeRuntime, out_fd: int) -> None:
         self.runtime = runtime
         self.out_fd = out_fd
         self.message_is_started = False
-        self.tool_calls: Optional[ToolCalls] = None
+        self.tool_calls_started = False
+        self.tool_calls = ToolCalls()
 
     async def handle_sse_object(self, sse_object: Mapping[str, Any]) -> None:
         delta = unwrap_delta(sse_object)
@@ -103,8 +74,9 @@ class SseHandler:
         tool_calls = delta.get("tool_calls")
         if tool_calls:
             assert isinstance(tool_calls, list), "Tool calls must be a list"
-            if self.tool_calls is None:
-                self.tool_calls = ToolCalls(tool_calls)
+            if not self.tool_calls_started:
+                self.tool_calls_started = True
+                self.tool_calls.extend(tool_calls)
             else:
                 self.tool_calls.delta(tool_calls)
